refactor: move location for merging\nhttps://stackoverflow.com/a/64228936/15887376
